requirements.txt
fastapi
pyTelegramBotAPI
uvicorn
python-dotenv